#171: create widgets-datetime-picker
