+++ publishToMavenLocal.sh
@@ -1,3 +1,3 @@
 ./gradlew -PpluginPublish publishPluginPublicationToMavenLocal
 ./gradlew -PlibraryPublish :widgets:publishToMavenLocal
-./gradlew -PlibraryPublish :widgets-flat:publishToMavenLocal :widgets-bottomsheet:publishToMavenLocal :widgets-sms:publishToMavenLocal
+./gradlew -PlibraryPublish :widgets-flat:publishToMavenLocal :widgets-bottomsheet:publishToMavenLocal :widgets-sms:publishToMavenLocal :widgets-datetime-picker:publishToMavenLocal
