+++ widgets-datetime-picker/src/androidMain/kotlin/dev/icerock/moko/widgets/datetimepicker/DatePickerDialogHandler.kt
@@ -2,7 +2,7 @@
  * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.moko.widgets.screen
+package dev.icerock.moko.widgets.datetimepicker
 
 import android.app.DatePickerDialog
 import android.app.Dialog
@@ -13,6 +13,7 @@ import androidx.fragment.app.FragmentManager
 import com.soywiz.klock.DateTime
 import dev.icerock.moko.graphics.Color
 import dev.icerock.moko.parcelize.Parcelize
+import dev.icerock.moko.widgets.screen.Screen
 import java.util.*
 import kotlin.properties.ReadOnlyProperty
