+++ widgets-datetime-picker/src/commonMain/kotlin/dev/icerock/moko/widgets/datetimepicker/ShowDatePickerExt.kt
@@ -2,10 +2,11 @@
  * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.moko.widgets.screen
+package dev.icerock.moko.widgets.datetimepicker
 
 import com.soywiz.klock.DateTime
 import dev.icerock.moko.graphics.Color
+import dev.icerock.moko.widgets.screen.Screen
 import kotlin.properties.ReadOnlyProperty
 
 expect class DatePickerDialogHandler
