+++ widgets-datetime-picker/src/commonMain/kotlin/dev/icerock/moko/widgets/datetimepicker/TimePickerDialog.kt
@@ -2,9 +2,10 @@
  * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.moko.widgets.screen
+package dev.icerock.moko.widgets.datetimepicker
 
 import dev.icerock.moko.graphics.Color
+import dev.icerock.moko.widgets.screen.Screen
 import kotlin.properties.ReadOnlyProperty
 
 expect class TimePickerDialogHandler
